+++ src/app/protected/test2/action.ts
@@ -1,0 +1,29 @@
+'use server'
+
+import { createClient } from '@supabase/supabase-js'
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL || ''
+const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || ''
+const supabase = createClient(supabaseUrl, supabaseKey)
+
+export async function fetchVideosWithTags() {
+  const { data, error } = await supabase.from('videos').select(`
+    *,
+    video_tags (
+      tags (
+        name
+      )
+    )
+  `)
+
+  if (error) {
+    throw new Error(error.message)
+  }
+
+  const formattedData = data.map((video) => ({
+    ...video,
+    video_tags: video.video_tags.map((tag: { tags: { name: string } }) => tag.tags.name),
+  }))
+
+  return formattedData
+}
